API [Partial] Provide serpentTools.ResultsFile
New results file object is provided that is responsible for
storing data from a result file, while the newly modified
ResultsReader is responsible for reading in data from a readable
stream (e.g. opened file), and populating resdata, universes,
and metadata dictionaries.

This a partial implementation as the full comparison interface
is yet to be added and tested.

The unit tests have been updated only as they pertain to
reading the files, with one exception. The result files have
no need to store the "variables" that were used in the processing
by passing variables and variableGroups to the class method. These
are used by the reader, and accurately, as the updated tests
indicate. However, I marked such tests as skipped for now until
decisions are made on how to test this, if at all.
